+++ src/main/java/org/example/repository/Midia.java
@@ -1,5 +1,5 @@
 package org.example.repository;
 public interface Midia {
-        int getClassificacao();
+        double getClassificacao();
 
 }
